+++ src/main/java/chess/domain/InitialPiece.java
@@ -7,6 +7,7 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
+import chess.domain.position.Position;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
