+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import java.util.List;
+import chess.domain.position.Position;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
