+++ src/test/java/chess/domain/MovablePathsTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import chess.domain.path.Direction;
 import chess.domain.path.MovablePaths;
 import chess.domain.path.Path;
+import chess.domain.position.Position;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
